requirements.txt
Flask==2.2.5
requests==2.28.1
python-dotenv==1.0.0
